requirements.txt
# Windows-Compatible Requirements
# No C++ build tools needed!

# Core CrewAI (without chromadb to avoid C++ compilation)
crewai==0.86.0
langchain-groq==0.2.1
python-dotenv==1.0.0
pydantic==2.9.2

# LangChain dependencies
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13

# HTTP and async support
requests==2.32.3
aiohttp==3.11.10